+++ dasbus/client/handler.py
@@ -27,8 +27,6 @@ from dasbus.signal import Signal
 from dasbus.constants import DBUS_FLAG_NONE
 from dasbus.specification import DBusSpecification
 from dasbus.typing import get_variant, get_variant_type, unwrap_variant
-from dasbus.typing import variant_replace_handles_with_fdlist_indices
-from dasbus.typing import variant_replace_fdlist_indices_with_handles
 
 import gi
 gi.require_version("Gio", "2.0")
@@ -37,7 +35,6 @@ from gi.repository import Gio, GLib
 
 __all__ = [
     "GLibClient",
-    "GLibClientUnix",
     "AbstractClientObjectHandler",
     "ClientObjectHandler"
 ]
@@ -168,80 +165,6 @@ class GLibClient(object):
             return message[len(prefix):]
 
         return message
-
-class GLibClientUnix(GLibClient):
-    """The low-level DBus client library based on GLib."""
-
-    @classmethod
-    def sync_call(cls, connection, service_name, object_path, interface_name,
-                  method_name, parameters, reply_type, flags=DBUS_FLAG_NONE,
-                  timeout=GLibClient.DBUS_TIMEOUT_NONE):
-        """Synchronously call a DBus method.
-
-        :return: a result of the DBus call
-        """
-        params = parameters
-        fds = None
-        if parameters:
-            params, fdlist = variant_replace_handles_with_fdlist_indices(
-                parameters)
-            if fdlist:
-                fds = Gio.UnixFDList.new_from_array(fdlist)
-
-        ret = connection.call_with_unix_fd_list_sync(
-            service_name,
-            object_path,
-            interface_name,
-            method_name,
-            params,
-            reply_type,
-            flags,
-            timeout,
-            fds,
-            None
-        )
-        return ret
-
-    @classmethod
-    def async_call(cls, connection, service_name, object_path, interface_name,
-                   method_name, parameters, reply_type, callback,
-                   callback_args=(), flags=DBUS_FLAG_NONE,
-                   timeout=GLibClient.DBUS_TIMEOUT_NONE):
-        """Asynchronously call a DBus method."""
-        params = parameters
-        fds = None
-        if parameters:
-            params, fdlist = variant_replace_handles_with_fdlist_indices(
-                parameters)
-
-            if fdlist:
-                fds = Gio.UnixFDList.new_from_array(fdlist)
-
-        connection.call_with_unix_fd_list(
-            service_name,
-            object_path,
-            interface_name,
-            method_name,
-            params,
-            reply_type,
-            flags,
-            timeout,
-            fds,
-            callback=cls._async_call_finish,
-            user_data=(callback, callback_args)
-        )
-
-    @classmethod
-    def _async_call_finish(cls, source_object, result_object, user_data):
-        """Finish an asynchronous DBus method call."""
-        # Prepare the user's callback.
-        callback, callback_args = user_data
-
-        # Call user's callback.
-        callback(
-            lambda: source_object.call_with_unix_fd_list_finish(result_object),
-            *callback_args
-        )
 
 
 class AbstractClientObjectHandler(metaclass=ABCMeta):
@@ -604,6 +527,11 @@ class ClientObjectHandler(AbstractClientObjectHandler):
         """
         # Unwrap a variant tuple.
         if fdlist is not None:
+            # pylint: disable=fixme,import-outside-toplevel
+            # FIXME: Move this code to the GLibClientUnix class.
+            from dasbus.unix import \
+                variant_replace_fdlist_indices_with_handles
+
             values = unwrap_variant(
                 variant_replace_fdlist_indices_with_handles(
                     result, fdlist.steal_fds()))
